move websocket connect LoginPage -> MainPage
