+++ src/pages/LoginPage.js
@@ -2,52 +2,15 @@ import styles from "../css/Account.module.css";
 import Button from "../css/Button.module.css";
 
 import { Link, useNavigate } from "react-router-dom";
-import { useState, useContext } from "react";
-import { SocketContext } from "../components/SocketContext.module";
 
 function LoginPage() {
-  const wsUrl = `ws://localhost/api/v2/websockets/ws`;
   const navigate = useNavigate();
-  const { setSocketResponse } = useContext(SocketContext);
-  const [socket, setSocket] = useState(null);
-
-  const getTokenAndCheck = async (storeID) => {
-    try {
-      const newSocket = new WebSocket(`${wsUrl}?s_id=${storeID}`);
-
-      newSocket.onopen = () => {
-        console.log("WebSocket connection established");
-      };
-
-      // 웹소켓 연결 시 받는 메시지
-      newSocket.onmessage = (event) => {
-        const response = JSON.parse(event.data);
-        console.log("Received response:", response);
-        if (response.type === "create_order") {
-          // console.log(response);
-          setSocketResponse(response);
-        }
-      };
-
-      // 웹소켓 연결 에러 발생 시
-      newSocket.onerror = (e) => {
-        console.log("WebSocket error:", e);
-      };
-
-      setSocket(newSocket);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
   const onLogin = async (e) => {
     e.preventDefault();
 
     const storeID = document.querySelector("#storeID").value;
     // storeID를 localStorage에 저장
     localStorage.setItem("storeID", storeID);
-
-    await getTokenAndCheck(storeID);
 
     navigate(`/main`);
   };
